api_yamdb/api/mixins.py
from rest_framework import mixins, viewsets


class ListCreateDestroyViewSet(
    mixins.ListModelMixin,
    mixins.CreateModelMixin,
    mixins.DestroyModelMixin,
    viewsets.GenericViewSet,
):
    """
    Creating a custom mixin with actions: List, Create, Destroy.
    """

    pass
